+++ component/repl-playground/vite.config.js
@@ -4,7 +4,7 @@ export default defineConfig({
   build: {
     rollupOptions: {
       external: (id) =>
-        /^(lit|@lit\/|prettier|@codemirror\/|codemirror|@replit\/codemirror-vim)/.test(id),
+        /^(lit|@lit\/|prettier|@codemirror\/|@lezer\/|codemirror|@replit\/codemirror-vim)/.test(id),
       output: {
         entryFileNames: `assets/[name].js`,
         chunkFileNames: `assets/[name].js`,