+++ site-entry.js
@@ -6,6 +6,7 @@ import '@jschofield/repl-playground';
 import '@jschofield/scroll-explain';
 import 'baseline-status';
 import './src/js/heading-links.js';
+import './src/js/theme-toggle.js';
 import MobileMenu from './src/js/mobile-menu.js';
 
 new MobileMenu('#mobile-menu');