+++ component/play-ground/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jschofield/play-ground
 
+## 0.3.0
+
+### Minor Changes
+
+- Added dark mode themes and styles
+
 ## 0.2.0
 
 ### Minor Changes